+++ src/supertux/statistics.cpp
@@ -139,7 +139,7 @@ Statistics::draw_worldmap_info(DrawingContext& context, float target_time)
   // check to see if screen size has been changed
   if (!(WMAP_INFO_TOP_Y1 == static_cast<float>(SCREEN_HEIGHT - 100))) {
     calculate_max_caption_length();
-    WMAP_INFO_LEFT_X = static_cast<float>(context.get_width() - 32 - m_max_width);
+    WMAP_INFO_LEFT_X = context.get_width() - 32.f - static_cast<float>(m_max_width);
     WMAP_INFO_RIGHT_X = WMAP_INFO_LEFT_X + static_cast<float>(m_max_width);
     WMAP_INFO_TOP_Y1 = static_cast<float>(SCREEN_HEIGHT - 100);
     WMAP_INFO_TOP_Y2 = WMAP_INFO_TOP_Y1 + 16;
@@ -212,12 +212,12 @@ Statistics::draw_endseq_panel(DrawingContext& context, Statistics* best_stats, c
 
   int box_w = 220+110+110;
   int box_h = 30+20+20+20;
-  int box_x = static_cast<int>((context.get_width() - box_w) / 2);
+  int box_x = static_cast<int>((static_cast<int>(context.get_width()) - box_w) / 2);
   int box_y = static_cast<int>(SCREEN_HEIGHT / 2) - box_h;
 
   int bd_w = static_cast<int>(backdrop->get_width());
   int bd_h = static_cast<int>(backdrop->get_height());
-  int bd_x = static_cast<int>((context.get_width() - bd_w) / 2);
+  int bd_x = static_cast<int>((static_cast<int>(context.get_width()) - bd_w) / 2);
   int bd_y = box_y + (box_h / 2) - (bd_h / 2);
 
   float col1_x = static_cast<float>(box_x);
